update search filter styles
