+++ wp-content/themes/labor-day/src/js/search/Presentational/Search Bar/SearchFilters.tsx
@@ -12,46 +12,59 @@ export default function SearchFilters({
 }) {
 	return (
 		<div className="cno-event-search-filters">
-			{filters.map(({ type: { name, filters } }) => (
-				<div className="cno-event-search-filters__container" key={name}>
-					<h4 className="cno-event-search-filters__title">{name}</h4>
-					<div className="cno-event-search-filters__filter-container">
-						<div className="dropdown">
-							<button
-								className="btn__outline--secondary dropdown-toggle"
-								type="button"
-								id={`${name}-dropdown`}
-								data-bs-toggle="dropdown"
-								aria-haspopup="true"
-								aria-expanded="false"
+			{filters.map((filter) => (
+				<SearchFilter
+					filter={filter}
+					key={filter.type.name}
+					dispatch={dispatch}
+					selectedFilters={selectedFilters}
+				/>
+			))}
+		</div>
+	);
+}
+
+function SearchFilter({ filter, dispatch, selectedFilters }) {
+	const {
+		type: { name, filters },
+	} = filter;
+	return (
+		<div className="cno-event-search-filters__container">
+			<div className="cno-event-search-filters__filter-container">
+				<div className="dropdown">
+					<button
+						className="btn__outline--secondary dropdown-toggle"
+						type="button"
+						id={`${name}-dropdown`}
+						data-bs-toggle="dropdown"
+						aria-haspopup="true"
+						aria-expanded="false"
+					>
+						{selectedFilters[name]}
+					</button>
+					<div
+						className="dropdown-menu"
+						aria-labelledby={`${name}-dropdown`}
+					>
+						{filters.map((filter, i) => (
+							<a
+								key={i}
+								role="button"
+								onClick={() =>
+									dispatch({
+										type: 'selectFilter',
+										payload: {
+											[name]: filter.name,
+										},
+									})
+								}
 							>
-								{selectedFilters[name]}
-							</button>
-							<div
-								className="dropdown-menu"
-								aria-labelledby={`${name}-dropdown`}
-							>
-								{filters.map((filter, i) => (
-									<a
-										key={i}
-										role="button"
-										onClick={() =>
-											dispatch({
-												type: 'selectFilter',
-												payload: {
-													[name]: filter.name,
-												},
-											})
-										}
-									>
-										{filter.name}
-									</a>
-								))}
-							</div>
-						</div>
+								{filter.name}
+							</a>
+						))}
 					</div>
 				</div>
-			))}
+			</div>
 		</div>
 	);
 }
