+++ wp-content/themes/labor-day/style.css
@@ -4,7 +4,7 @@ Theme URI: https://choctawnation.com/
 Author: Choctaw Nation of Oklahoma
 Author URI: https://choctawnation.com/
 Description: A new theme for Choctaw Labor Day Festival.
-Version: 1.5.5
+Version: 1.5.6
 Tested up to: 6.2.2
 Requires PHP: 8.0.22
 */
